Added paginated presenter
